Fixing Linux headless makefile
The makefile now creates the src/obj directory if it doesn't
exist. It also fixes the references to the files in the src/hash
directory, which the makefile previously looked for in src/obj.

Additionally, the clean command was updated to remove object files
from src/hash and the entire src/obj directory.