Fixing wallet balance display issues
The wallet never updated its balance because the
updateDisplayUnit() method never passed it updated balance or
stake information. The method now grabs this data directly from
the wallet, as well as the immature coin balance, and passes them
to an updated setBalance method, which correctly updates the
display after each block is accepted.