Fixing PoS (hardfork)
The main issue with the previous PoS methods were the calculation
of bnCoinDayWeight (which always returned incorrectly during
calculation) and GetKernelStakeModifier(), which would never
return true. It turns out that the stake modifier interval was
far too slow to outpace the growth of the block time, so PoS would
never occur. This commit fixes bnCoinDayWeight and lowers the
stake modifier interval (to once every 30 minutes).

The additional commits revert the usage of "V2" algorithms back to
"V1" algorithms for Weight and Network Weight calculations, as
the "V2" algorithms are unsuitable for actual calculations.