Fixing some issues with staking display
Previously, the staking display would feed back wildly inaccurate
values for coin weight and occasionally decide that obviously
mature coins were immature. This commit reverts the TruckCoin
update that was recently applied, removes the zeroing of nWeight,
and adds an if statement to choose between target stake spacing
intervals, which are varied in the code based on block height.