Fixing PoS reward check; fixing blockchain syncing
The PoS reward incorrectly compared FORK_POS_REWARD_CHANGE to
nTime (block time) rather than nHeight (block height) previously.
Because it was done for a hard-fork, the difference that this
made was negligible, but it was still incorrect.

Additionally, the version 2.3.0.0 commits removed the hard
limit on PoS rewards without wrapping it behind a fork. This
update restores the hard limit because all blocks before the fork
ended up using this reward anyway.

Blockchain syncing previously paused regularly when trying to
sync from a blockchain with less than 2500 blocks (the end of the
PoW phase). This was because the stake modifier interval was
changed, but no traces of the modifier interval for the first
2500 blocks was accessible by the wallet code. By moving the
code for checking incoming PoS blocks into AcceptBlock(), we
are able to leverage nHeight and pass it into
CheckProofOfStake(). This allows the wallet to select between
the two stake modifier intervals when processing all blocks.