Redirecting IRC to new Rizon channel
The in-wallet IRC has not worked up to this point because it was
directed to Freenode, where the #deoxyribose channel did not
exist. A new #deoxyribose channel has been created at
irc.rizon.net. This commit redirects the wallet to Rizon's
server and allows the client to function as intended.