Fixing some userlist bugs, adding CTCP VERSION response
The user list in the IRC client has been slightly buggy since
the wallet came to be. This commit adds checks to make sure
we don't print End of List messages, or phrases that
start with http:, https:, or Rizon's message header flag in the
user list.

Additionally, the update adds support for the /privmsg command,
which was conspicuously absent from the wallet.

The cpp and header files for src/chatwindow.* and
src/serveur.* were also cleaned up to follow a 4-space
indentation per tab pattern because they were an eyesore.
Brackets have been sprinkled about for consistency as well.

Last, to conform with Rizon's rules, the wallet's IRC client
needed to support the CTCP VERSION command. This commit
implements a basic version of that command (in a roundabout
way, but false positives are better than accidental bans until
a better solution is discovered).